"Morse code" normalisation with footway symbols
This seems to make more intuitive sense: since footways are dotted
paths, the ped-sharing should be dot-dot-dash.  And that leaves plain
dashes as the clearest other symbol for shared-with-limited-traffic.

I'd have loved to make it an indictment of the act of sharing with ANY
traffic, but I think this will lead to the least confusion overall.